+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+-   Added `baseDirectory` option
+
 ## [2.0.0] - 2019-04-27
 
 -   Add named export: `exConfig`